Change integer to float for base and height

+++ Area do triangulo.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main(int argc, char** argv)
 {
 
-	int base = 0, altura = 0;
+	float base = 0, altura = 0;
 	cout << "##### AREA DO TRIANGULO #####\n\n >>> (base x altura)/2 <<<\n\n";
 	
 	cout << "Informe o valor da base (em metros)? -> ";
@@ -12,7 +12,7 @@ int main(int argc, char** argv)
 	cout << "Informe o valor da altura (em metros)? ->";
 	cin >> altura;
 
-	/*Exibindo o valor da área calculada*/	
+	/*Exibindo o valor da Ã¡rea calculada*/	
 	cout << "\n Area do Triangulo -> ("<<base<<"m x "<<altura<<"m)/2 = "<< ((base * altura)/2)<< "m2";
 	return 0;
 }
